+++ ididi/_type_resolve.py
@@ -255,7 +255,7 @@ def flatten_annotated(typ: Annotated[Any, Any]) -> list[Any]:
 
 
 @lru_cache(CacheMax)
-def get_bases(dependent: Union[type, GenericAlias]) -> tuple[type, ...]:
+def get_bases(dependent: Union[type, GenericAlias]) -> tuple[Union[type, GenericAlias], ...]:
     if not isinstance(dependent, type):
         return (dependent,)